postgres_to_es/requirements/base.txt
# основные пакеты
psycopg2-binary==2.9.1
requests==2.26.0
pydantic==1.8.2


# зависимости
asgiref==3.4.1
pytz==2021.1
sqlparse==0.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
urllib3==1.26.7
typing-extensions==3.10.0.2
